chore: relocate internal package used for testing.
