+++ instrumentation/opentelemetry/internal/tracetesting/recorder.go
@@ -1,4 +1,4 @@
-package internal
+package tracetesting
 
 import (
 	"context"
@@ -6,27 +6,27 @@ import (
 	"go.opentelemetry.io/otel/sdk/trace"
 )
 
-var _ trace.SpanExporter = &Recorder{}
+var _ trace.SpanExporter = &recorder{}
 
-// Recorder records spans being synced through the SpanSyncer interface.
-type Recorder struct {
+// recorder records spans being synced through the SpanSyncer interface.
+type recorder struct {
 	spans []trace.ReadOnlySpan
 }
 
 // ExportSpans records spans into the internal buffer
-func (r *Recorder) ExportSpans(ctx context.Context, s []trace.ReadOnlySpan) error {
+func (r *recorder) ExportSpans(_ context.Context, s []trace.ReadOnlySpan) error {
 	r.spans = append(r.spans, s...)
 	return nil
 }
 
 // Shutdown flushes the buffer
-func (r *Recorder) Shutdown(_ context.Context) error {
+func (r *recorder) Shutdown(_ context.Context) error {
 	_ = r.Flush()
 	return nil
 }
 
 // Flush returns the current recorded spans and reset the recordings
-func (r *Recorder) Flush() []trace.ReadOnlySpan {
+func (r *recorder) Flush() []trace.ReadOnlySpan {
 	spans := r.spans
 	r.spans = nil
 	return spans
