+++ instrumentation/opentelemetry/internal/tracetesting/spanattributes_test.go
@@ -1,4 +1,4 @@
-package internal
+package tracetesting
 
 import (
 	"testing"
